Consolidate AMQP connection integration tests.
